Moved everything related to Chord dialog to a separate directory ui/chord
