+++ kguitar/ui/chord/chordanalyzer.cpp
