+++ kguitar/ui/chord/chordanalyzer.h
