+++ kguitar/ui/chord/chordlistitem.cpp
