+++ kguitar/ui/chord/fingering.cpp
@@ -1,4 +1,4 @@
-#include "fingers.h"
+#include "fingering.h"
 #include "data/tabtrack.h"
 #include "settings.h"
