+++ kguitar/ui/chord/fingerlist.cpp
