+++ kguitar/ui/chord/strumming.cpp
